Check for explicit true value from result
Make sure to check for an explicit `true` value for the result of the property
function.  The result value may be truthy, but in fact contain info about an
error encountered in the property function.  Without this, a test might pass
when an error is thrown within the property function.